+++ src/scripts/directives.js
@@ -152,10 +152,16 @@
                 },
                 templateUrl: 'templates/type/radio-input.html',
                 link: function (scope) {
-                    scope.localModel = {
-                        value: scope.model
-                    };
-
+                    scope.localModel = '';
+                    var modelWatchDeregister = scope.$watch('model', function (value) {
+                        if (angular.isDefined(value)) {
+                            scope.localModel = {
+                                value: scope.model
+                            };
+                            // Deregister the watch to avoid having a loop with the $watch below
+                            modelWatchDeregister();
+                        }
+                    });
                     scope.$watch('localModel.value', function (value) {
                         scope.model = value;
                     });